requirements.txt
annotated-types==0.7.0
anyio==4.13.0
attrs==26.1.0
bcrypt==5.0.0
certifi==2026.2.25
cffi==2.0.0
click==8.3.2
cryptography==46.0.7
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.3
idna==3.11
invoke==3.0.3
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
markdown-it-py==4.0.0
mcp==1.27.0
mdurl==0.1.2
netmiko==4.6.0
ntc_templates==9.1.0
paramiko==4.0.0
pycparser==3.0
pydantic==2.12.5
pydantic-settings==2.13.1
pydantic_core==2.41.5
Pygments==2.20.0
PyJWT==2.12.1
PyNaCl==1.6.2
pyserial==3.5
python-dotenv==1.2.2
python-multipart==0.0.24
PyYAML==6.0.3
referencing==0.37.0
rich==14.3.3
rpds-py==0.30.0
ruamel.yaml==0.19.1
scp==0.15.0
sse-starlette==3.3.4
starlette==1.0.0
textfsm==2.1.0
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.44.0
